Split ParseResult into separate file.
